Alterado gestor de dependências para pipenv
close #19
