+++ .github/workflows/.githubactions.yml
@@ -8,23 +8,23 @@ jobs:
       - name: Configurar Python 3.10
         uses: actions/setup-python@v2
         with:
-          python-version: '3.10'
+          python-version: "3.10"
       - name: Instalar Dependências
         run: |
           python -m pip install --upgrade pip
-          pip install flake8 pytest
-          pip install codecov
-          pip install -q -r requirements-dev.txt
-      - name: Rodar Flake8
+          pip install -q pipenv codecov
+          pipenv sync --dev
+      - name: Rodar com flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names 
-          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics 
+          pipenv run flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics 
           # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide 
-          flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+          pipenv run flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Rodar Testes Automáticos com Pytest
         run: |
-          pytest --cov=libpythonpro
-      - name: Cobertura de Testes com Codecov
-        uses: codecov/codecov-action@v3
-        with:
-          token: ${{secrets.CODECOV_TOKEN}}
+          pipenv run pytest --cov=libpythonpro
+      - name: Relatório Coverage
+        env:
+          CODECOV_TOKEN: ${{secrets.CODECOV_TOKEN}}
+        run: |
+          pipenv run codecov
