+++ libpythonpro/github_api.py
@@ -10,4 +10,3 @@ def buscar_avatar(usuario):
     url = f'https://api.github.com/users/{usuario}'
     resp = requests.get(url)
     return resp.json()['avatar_url']
-
