+++ libpythonpro/spam/modelos.py
@@ -1,4 +1,5 @@
 class Usuario:
-    def __init__(self, nome):
+    def __init__(self, nome, email):
+        self.email = email
         self.nome = nome
         self.id = None
