--- libpythonpro/test_spam/test_usuarios.py
@@ -1,34 +1,0 @@
-import pytest
-
-from libpythonpro.spam.db import Conexao
-from libpythonpro.spam.modelos import Usuario
-
-
-@pytest.fixture
-def conexao():
-    #Setup
-    conexao_obj = Conexao()
-    yield conexao_obj
-    #Tear Down
-    conexao_obj.fechar()
-
-
-@pytest.fixture
-def sessao(conexao):
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
-
-
-def test_salvar_usuario(sessao):
-    usuario = Usuario(nome='Daniel')
-    sessao.salvar(usuario)
-    assert isinstance(usuario.id, int)
-
-
-def test_listar_usuarios(sessao):
-    usuarios = [Usuario(nome='Daniel'), Usuario(nome='Geraldo')]
-    for usuario in usuarios:
-        sessao.salvar(usuario)
-    assert usuarios == sessao.listar()
