+++ libpythonpro/tests/test_spam/test_envio_para_base_de_usuarios.py
@@ -6,6 +6,7 @@ from libpythonpro.spam.enviador_de_email import Enviador
 from libpythonpro.spam.main import EnviadordeSpam
 from libpythonpro.spam.modelos import Usuario
 
+
 @pytest.mark.parametrize(
     'usuarios',
     [
