+++ libpythonpro/tests/test_spam/test_usuarios.py
@@ -1,0 +1,17 @@
+from libpythonpro.spam.modelos import Usuario
+
+
+def test_salvar_usuario(sessao):
+    usuario = Usuario(nome='Daniel', email='danielrodriguesr@gmail.com')
+    sessao.salvar(usuario)
+    assert isinstance(usuario.id, int)
+
+
+def test_listar_usuarios(sessao):
+    usuarios = [
+        Usuario(nome='Daniel', email='danielrodriguesr@gmail.com'),
+        Usuario(nome='Geraldo', email='jgrsilva55@gmail.com')
+    ]
+    for usuario in usuarios:
+        sessao.salvar(usuario)
+    assert usuarios == sessao.listar()
