--- requirements-dev.txt
@@ -1,25 +1,0 @@
-# Deps do Flake8
-flake8==5.0.4
-mccabe==0.7.0
-pycodestyle==2.9.1
-pyflakes==2.5.0
-
-# Deps do Pytest
-attrs==22.1.0
-colorama==0.4.5
-iniconfig==1.1.1
-packaging==21.3
-pluggy==1.0.0
-py==1.11.0
-pyparsing==3.0.9
-pytest==7.1.3
-tomli==2.0.1
-
-# Deps do Coverage
-coverage==6.5.0
-pytest-cov==4.0.0
-
-# Deps do Mock
-pytest-mock==3.9.0
-
--r requirements.txt
